refactors tests into test/test_optim.py
instead of test/optim/test_optim.py
