+++ test/optim/test_optim.py
@@ -7,7 +7,6 @@ from typing import Any
 import torch
 from torch import nn, Tensor
 from torch.optim import Optimizer, SGD
-from torch.testing._internal.common_optimizers import optim_db, OptimizerInfo, optims
 from torch.testing._internal.common_utils import (
     gradcheck,
     load_tests,
@@ -19,27 +18,6 @@ from torch.testing._internal.common_utils import (
 # load_tests from common_utils is used to automatically filter tests for
 # sharding on sandcastle. This line silences flake warnings
 load_tests = load_tests
-
-
-def _diff_fn(p, grad, opt_differentiable_state, opt_class, kwargs, *ignored):
-    # Ignored is the list of values in `opt_differentiable_state`, we do this
-    # for `gradcheck` to correctly track the state tensors as function inputs
-    # because otherwise it can't unpack the values in the `opt_differentiable_state`
-    # dict
-    p = p.clone()
-    p.grad = grad
-    opt_differentiable_state = {
-        k: v.clone() if isinstance(v, torch.Tensor) else v
-        for k, v in opt_differentiable_state.items()
-    }
-    opt = opt_class([p], **kwargs)
-    opt.state[p].update(opt_differentiable_state)
-    opt.step()
-    return (p,) + tuple(
-        v
-        for v in opt.state[p].values()
-        if isinstance(v, torch.Tensor) and v.requires_grad
-    )
 
 
 def _multistep_backprop_diff_hyperparams_fn(
@@ -115,64 +93,6 @@ def _multistep_backprop_diff_hyperparams_fn(
         )
         + tuple(differentiable_kwargs)
     )
-
-
-def _get_optimizer_state(
-    optim_cls: type[Optimizer],
-    params: Tensor,
-    optim_kwargs: dict[str, Any],
-) -> dict[str, Any]:
-    params.grad = torch.rand_like(params)
-
-    optim = optim_cls([params], **optim_kwargs)
-    optim.step()
-
-    states = optim.state_dict()["state"]
-    # Accesses state of the first (and only) parameter. When using SGD w/ no momentum, there is no state
-    state = states[0] if 0 in states else states
-    return state
-
-
-class TestDifferentiableOptimizerOptimizerInfo(TestCase):
-    @optims(optim_db, dtypes=[torch.float64])
-    def test_optimizers_differentiable_wrt_params(
-        self, device: str, dtype: torch.dtype, optim_info: OptimizerInfo
-    ):
-        optim_cls = optim_info.optim_cls
-        if "differentiable" not in optim_info.supported_impls:
-            self.skipTest(f"Differentiable {optim_cls.__name__} not supported")
-
-        for optim_input in optim_info.optim_inputs_func(device=device):
-            mock_params = torch.rand(10, requires_grad=True, device=device, dtype=dtype)
-            mock_optim_kwargs = optim_input.kwargs.copy()
-            optim_state = _get_optimizer_state(
-                optim_cls, mock_params, mock_optim_kwargs
-            )
-
-            optim_kwargs = optim_input.kwargs
-            optim_kwargs.update({"differentiable": True})
-
-            # Cannot be capturable and differentiable
-            if "capturable" in optim_kwargs:
-                optim_kwargs.update({"capturable": False})
-
-            # sgd defaults to foreach on cuda so must explicitly specify not foreach
-            if optim_cls is SGD and device.startswith("cuda"):
-                optim_kwargs.update({"foreach": False})
-
-            params = torch.rand(10, requires_grad=True, device=device, dtype=dtype)
-            grad = torch.rand(10, requires_grad=True, device=device, dtype=dtype)
-            gradcheck(
-                _diff_fn,
-                (
-                    params,
-                    grad,
-                    optim_state,
-                    optim_cls,
-                    optim_kwargs,
-                    *optim_state.values(),
-                ),
-            )
 
 
 @skipIfTorchDynamo("Differentiable optimizers not supported")
@@ -252,6 +172,4 @@ class TestDifferentiableOptimizer(TestCase):
 
 
 if __name__ == "__main__":
-    A = TestDifferentiableOptimizerOptimizerInfo()
-    A.test_optimizers_differentiable_wrt_params("cpu", torch.float64, optim_db[-2])
     print("These tests should be run through test/test_optim.py instead")
