+++ test/test_optim.py
@@ -8,10 +8,7 @@ from typing import Any, Dict, Tuple
 from unittest.mock import patch
 
 from optim.test_lrscheduler import TestLRScheduler  # noqa: F401
-from optim.test_optim import (  # noqa: F401
-    TestDifferentiableOptimizer,
-    TestDifferentiableOptimizerOptimizerInfo,
-)
+from optim.test_optim import TestDifferentiableOptimizer  # noqa: F401
 from optim.test_swa_utils import TestSWAUtils  # noqa: F401
 
 import torch
@@ -38,10 +35,12 @@ from torch.testing._internal.common_optimizers import (
     _get_optim_inputs_including_global_cliquey_kwargs,
     optim_db,
     OptimizerErrorEnum,
+    OptimizerInfo,
     optims,
     TensorTracker,
 )
 from torch.testing._internal.common_utils import (
+    gradcheck,
     markDynamoStrictTest,
     parametrize,
     run_tests,
@@ -2219,11 +2218,90 @@ class TestOptimRenewed(TestCase):
             for state in optim.state.values():
                 self.assertGreater(len(state), 0)
 
+    @optims(optim_db, dtypes=[torch.float64])
+    def test_optimizers_differentiable_wrt_params(
+        self, device: str, dtype: torch.dtype, optim_info: OptimizerInfo
+    ):
+        def _diff_fn(
+            p: torch.Tensor,
+            grad: torch.Tensor,
+            opt_differentiable_state: dict[str, Any],
+            opt_class: type[Optimizer],
+            kwargs: dict[str, Any],
+            *ignored,
+        ) -> tuple[torch.Tensor, ...]:
+            # Ignored is the list of values in `opt_differentiable_state`, we do this
+            # for `gradcheck` to correctly track the state tensors as function inputs
+            # because otherwise it can't unpack the values in the `opt_differentiable_state`
+            # dict
+            p = p.clone()
+            p.grad = grad
+            opt_differentiable_state = {
+                k: v.clone() if isinstance(v, torch.Tensor) else v
+                for k, v in opt_differentiable_state.items()
+            }
+            opt = opt_class([p], **kwargs)
+            opt.state[p].update(opt_differentiable_state)
+            opt.step()
+            return (p,) + tuple(
+                v
+                for v in opt.state[p].values()
+                if isinstance(v, torch.Tensor) and v.requires_grad
+            )
+
+        def _get_optimizer_state(
+            optim_cls: type[Optimizer],
+            params: torch.Tensor,
+            optim_kwargs: dict[str, Any],
+        ) -> dict[str, Any]:
+            params.grad = torch.rand_like(params)
+
+            optim = optim_cls([params], **optim_kwargs)
+            optim.step()
+
+            states = optim.state_dict()["state"]
+            # Accesses state of the first (and only) parameter. When using SGD w/ no momentum, there is no state
+            state = states[0] if 0 in states else states
+            return state
+
+        optim_cls = optim_info.optim_cls
+        if "differentiable" not in optim_info.supported_impls:
+            self.skipTest(f"Differentiable {optim_cls.__name__} not supported")
+
+        for optim_input in optim_info.optim_inputs_func(device=device):
+            mock_params = torch.rand(10, requires_grad=True, device=device, dtype=dtype)
+            mock_optim_kwargs = optim_input.kwargs.copy()
+            optim_state = _get_optimizer_state(
+                optim_cls, mock_params, mock_optim_kwargs
+            )
+
+            optim_kwargs = optim_input.kwargs
+            optim_kwargs.update({"differentiable": True})
+
+            # Cannot be capturable and differentiable
+            if "capturable" in optim_kwargs:
+                optim_kwargs.update({"capturable": False})
+
+            # sgd defaults to foreach on cuda so must explicitly specify not foreach
+            if optim_cls is SGD and device.startswith("cuda"):
+                optim_kwargs.update({"foreach": False})
+
+            params = torch.rand(10, requires_grad=True, device=device, dtype=dtype)
+            grad = torch.rand(10, requires_grad=True, device=device, dtype=dtype)
+            gradcheck(
+                _diff_fn,
+                (
+                    params,
+                    grad,
+                    optim_state,
+                    optim_cls,
+                    optim_kwargs,
+                    *optim_state.values(),
+                ),
+            )
+
 
 instantiate_device_type_tests(TestOptimRenewed, globals(), allow_mps=True)
-instantiate_device_type_tests(
-    TestDifferentiableOptimizerOptimizerInfo, globals(), allow_mps=True
-)
 
 if __name__ == "__main__":
     run_tests()
